Add customHooks
