+++ src/App.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react'
+import React from 'react'
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 import ContactList from './components/ContactList/ContactList'
 import HomePage from './components/HomePage/HomePage'
@@ -7,37 +7,26 @@ import UserProfileForm from './components/UserProfile/UserProfileForm'
 import CreateContact from './components/CreateContactForm/CreateContact'
 import NoContact from './components/ContactList/NoContact'
 import NoFavorite from './components/ContactList/NoFavorite'
-
 import ContactProfile from './components/ContactProfile/ContactProfile'
 import FavContactList from './components/ContactList/FavContactList'
+import useContacts from './hooks/useContacts'
+import useUser from './hooks/useUser'
+import useFavorites from './hooks/useFavorites'
 
 export default function App() {
-  const [contacts, setContacts] = useState(
-    JSON.parse(localStorage.getItem('savedContact')) || []
-  )
-
-  const [userData, setUserData] = useState(
-    JSON.parse(localStorage.getItem('savedUser')) || ['']
-  )
-  const user = userData[0]
-
-  const [favorites, setFavorites] = useState(
-    JSON.parse(localStorage.getItem('favorites')) || []
-  )
-
-  useEffect(() => {
-    saveLocally('favorites', favorites)
-  }, [favorites])
+  const [contacts, addContact] = useContacts()
+  const [savedUser, addUser] = useUser()
+  const [favorites, toggleFavorite] = useFavorites()
 
   return (
     <Router>
       <Switch>
         <Route exact path="/">
-          <HomePage user={user} />
+          <HomePage user={savedUser} />
         </Route>
         <Route path="/user-profile">
-          {localStorage.getItem('savedUser') !== null ? (
-            <UserProfile user={user} />
+          {savedUser ? (
+            <UserProfile user={savedUser} />
           ) : (
             <UserProfileForm onSubmit={addUser} />
           )}
@@ -46,7 +35,7 @@ export default function App() {
           <UserProfileForm onSubmit={addUser} />
         </Route>
         <Route path="/favorites">
-          {localStorage.getItem('favorites') !== null ? (
+          {favorites.length ? (
             <FavContactList favorites={favorites} />
           ) : (
             <NoFavorite />
@@ -56,7 +45,7 @@ export default function App() {
           <CreateContact onSubmit={addContact} />
         </Route>
         <Route path="/list">
-          {localStorage.getItem('savedContact') !== null ? (
+          {contacts.length ? (
             <ContactList contacts={contacts} />
           ) : (
             <NoContact />
@@ -72,34 +61,4 @@ export default function App() {
       </Switch>
     </Router>
   )
-
-  function addContact(contactItem) {
-    setContacts([...contacts, contactItem])
-    localStorage.setItem(
-      'savedContact',
-      JSON.stringify([...contacts, contactItem])
-    )
-  }
-  function addUser(userItem) {
-    setUserData([userItem])
-    localStorage.setItem('savedUser', JSON.stringify([userItem]))
-  }
-
-
-  function toggleFavorite(favorite) {
-    const index = favorites.findIndex(
-      (favoriteItem) => favoriteItem.id === favorite.id
-    )
-    index > -1
-      ? setFavorites([
-          ...favorites.slice(0, index),
-          ...favorites.slice(index + 1),
-        ])
-      : setFavorites([...favorites, { ...favorite }])
-  }
-
-  function saveLocally(favorites) {
-    setUserData([favorites])
-    localStorage.setItem('favorites', JSON.stringify(favorites))
-  }
 }
